+++ src/main.js
@@ -19,7 +19,12 @@ import { validatedSchemaModel} from './schemaModelValidator.js';
 import { getNeptuneSchema, setGetNeptuneSchemaParameters } from './NeptuneSchema.js';
 import { createUpdateAWSpipeline, removeAWSpipelineResources } from './pipelineResources.js'
 import { createAWSpipelineCDK } from './CDKPipelineApp.js'
-import { createApolloDeploymentPackage, createLambdaDeploymentPackage, getModulePath } from './zipPackage.js'
+import {
+    createApolloDeploymentPackage,
+    createLambdaDeploymentPackage,
+    getModulePath,
+    compressToGzipFile
+} from './zipPackage.js'
 import { loggerDebug, loggerError, loggerInfo, loggerInit, yellow } from './logger.js';
 
 import ora from 'ora';
@@ -513,7 +518,8 @@ async function main() {
     if (inputGraphQLSchemaChanges != '') {
         inputGraphQLSchema = changeGraphQLSchema(inputGraphQLSchema, inputGraphQLSchemaChanges); 
     }
-    
+
+    const resolverSchemaFile = path.join(outputFolderPath, `${outputFilePrefix}.resolver.schema.json.gz`);
     if (inputGraphQLSchema != '') {
         // Parse schema
         schemaModel = schemaParser(inputGraphQLSchema);
@@ -523,11 +529,11 @@ async function main() {
 
         // Generate schema for resolver
         const queryDataModelJSON = JSON.stringify(schemaModel, null, 2);
-        const resolverSchemaFile = path.join(outputFolderPath, `${outputFilePrefix}.resolver.schema.json`);
 
         try {
-            writeFileSync(resolverSchemaFile, queryDataModelJSON);
-            loggerInfo('Wrote schema for resolver to file: ' + yellow(resolverSchemaFile), {toConsole: true});
+            // resolver schema is compressed for better schema initialization performance
+            await compressToGzipFile(queryDataModelJSON, resolverSchemaFile);
+            loggerInfo('Wrote compressed schema for resolver to file: ' + yellow(resolverSchemaFile), {toConsole: true});
         } catch (err) {
             loggerError('Error writing schema for resolver to file: ' + yellow(resolverSchemaFile), err);
         }
@@ -623,11 +629,10 @@ async function main() {
                 }
             break;
         }
-
+        
         // output Apollo zip
         if (createUpdateApolloServer || createUpdateApolloServerSubgraph) {
             const apolloZipPath = path.join(outputFolderPath, `apollo-server-${neptuneInfo.graphName}-${new Date().getTime()}.zip`);
-            const resolverSchemaFilePath = path.join(outputFolderPath, `${outputFilePrefix}.resolver.schema.json`);
             try {
                 if (!quiet) {
                     spinner = ora('Creating Apollo server ZIP file ...').start();
@@ -636,7 +641,7 @@ async function main() {
                     zipFilePath: apolloZipPath,
                     resolverFilePath: outputLambdaResolverFile,
                     schemaFilePath: outputSchemaFile,
-                    resolverSchemaFilePath: resolverSchemaFilePath,
+                    resolverSchemaFilePath: resolverSchemaFile,
                     neptuneInfo: neptuneInfo,
                     isSubgraph: createUpdateApolloServerSubgraph
                 });
@@ -665,7 +670,8 @@ async function main() {
                 await createLambdaDeploymentPackage({
                     outputZipFilePath: outputLambdaResolverZipFile,
                     templateFolderPath: path.join(__dirname, outputLambdaPackagePath),
-                    resolverFilePath: outputLambdaResolverFile
+                    resolverFilePath: outputLambdaResolverFile,
+                    resolverSchemaFilePath: resolverSchemaFile
                 });
                 if (!quiet) {
                     spinner.succeed('Created Lambda ZIP');