+++ templates/Lambda4AppSyncGraphSDK/index.mjs
@@ -1,12 +1,16 @@
-import {ExecuteQueryCommand, NeptuneGraphClient} from "@aws-sdk/client-neptune-graph";
-import {resolveGraphDBQueryFromAppSyncEvent, initSchema} from './output.resolver.graphql.js';
-import {readFileSync} from "fs";
+import { ExecuteQueryCommand, NeptuneGraphClient } from "@aws-sdk/client-neptune-graph";
+import { initSchema, resolveGraphDBQueryFromAppSyncEvent } from './output.resolver.graphql.js';
+import { decompressGzipToString } from './util.mjs';
 
 const PROTOCOL = 'https';
 const QUERY_LANGUAGE = 'OPEN_CYPHER';
 const RESOLVER_LANGUAGE = 'opencypher';
 
 let client;
+
+const schemaDataModelJSON = await decompressGzipToString('output.resolver.schema.json.gz');
+const schemaModel = JSON.parse(schemaDataModelJSON);
+initSchema(schemaModel);
 
 function getClient() {
     if (!client) {
@@ -50,9 +54,6 @@ function log(message) {
  */
 function resolveGraphQuery(event) {
     try {
-        const schemaDataModelJSON = readFileSync('output.resolver.schema.json', 'utf-8');
-        let schemaModel = JSON.parse(schemaDataModelJSON);
-        initSchema(schemaModel);
         let resolver = resolveGraphDBQueryFromAppSyncEvent(event);
         if (resolver.language !== RESOLVER_LANGUAGE) {
             return onError('Unsupported resolver language:' + resolver.language)