+++ templates/Lambda4AppSyncSDK/index.mjs
@@ -1,12 +1,24 @@
-import { NeptunedataClient, ExecuteOpenCypherQueryCommand, ExecuteGremlinQueryCommand } from "@aws-sdk/client-neptunedata";
-import {resolveGraphDBQueryFromAppSyncEvent, refactorGremlinqueryOutput, initSchema} from './output.resolver.graphql.js';
-import {readFileSync} from "fs";
+import {
+    ExecuteGremlinQueryCommand,
+    ExecuteOpenCypherQueryCommand,
+    NeptunedataClient
+} from "@aws-sdk/client-neptunedata";
+import {
+    initSchema,
+    refactorGremlinqueryOutput,
+    resolveGraphDBQueryFromAppSyncEvent
+} from './output.resolver.graphql.js';
+import { decompressGzipToString } from './util.mjs';
 
 const LOGGING_ENABLED = process.env.LOGGING_ENABLED;
 
 const config = {            
     endpoint: `https://${process.env.NEPTUNE_HOST}:${process.env.NEPTUNE_PORT}`
 };
+
+const schemaDataModelJSON = await decompressGzipToString('output.resolver.schema.json.gz');
+const schemaModel = JSON.parse(schemaDataModelJSON);
+initSchema(schemaModel);
 
 let client;
 
@@ -42,9 +54,6 @@ export const handler = async(event) => {
     let resolver = { query:'', parameters:{}, language: 'opencypher', fieldsAlias: {} };
 
     try {
-        const schemaDataModelJSON = readFileSync('output.resolver.schema.json', 'utf-8');
-        let schemaModel = JSON.parse(schemaDataModelJSON);
-        initSchema(schemaModel);
         resolver = resolveGraphDBQueryFromAppSyncEvent(event);
         if (LOGGING_ENABLED) console.log(JSON.stringify(resolver, null, 2));
     } catch (error) {