+++ test/TestCases/Case01/Case01.03.test.js
@@ -6,7 +6,7 @@ const casetest = readJSONFile('./test/TestCases/Case01/case.json');
 try {
     await testResolverQueriesResults(   './TestCases/Case01/output/output.resolver.graphql.js',
         './test/TestCases/Case01/queries',
-        './test/TestCases/Case01/output/output.resolver.schema.json',
+        './test/TestCases/Case01/output/output.resolver.schema.json.gz',
         casetest.host,
         casetest.port);
 } finally {