+++ test/TestCases/Case02/Case02.02.test.js
@@ -12,7 +12,7 @@ describe('Validate output content', () => {
         'output.jsresolver.graphql.js',
         'output.neptune.schema.json',
         'output.resolver.graphql.js',
-        'output.resolver.schema.json',
+        'output.resolver.schema.json.gz',
         'output.schema.graphql',
         'output.source.schema.graphql'
     ]);