+++ test/TestCases/Case07/Case07.02.test.js
@@ -1,4 +1,4 @@
-import { checkFolderContainsFiles, compareFileContents, unzipAndGetContents } from '../../testLib';
+import { checkFolderContainsFiles, compareFileContents, executeAppSyncQuery, unzipAndGetContents } from '../../testLib';
 import path from "path";
 import fs from "fs";
 
@@ -7,7 +7,7 @@ const referenceFolder = './test/TestCases/Case07/outputReference';
 
 describe('Validate pipeline with sdk resolver output content', () => {
     const expectedFiles = [
-        'AirportsJestSDKTest.resolver.schema.json',
+        'AirportsJestSDKTest.resolver.schema.json.gz',
         'AirportsJestSDKTest.zip',
         'AirportsJestSDKTest-resources.json',
         'sdk.resolver.graphql.js',
@@ -28,9 +28,10 @@ describe('Validate pipeline with sdk resolver output content', () => {
             'index.mjs',
             'node_modules',
             'output.resolver.graphql.js',
-            'output.resolver.schema.json',
+            'output.resolver.schema.json.gz',
             'package-lock.json',
-            'package.json'
+            'package.json',
+            'util.mjs'
         ];
 
 
@@ -42,4 +43,13 @@ describe('Validate pipeline with sdk resolver output content', () => {
         const fileContent = fs.readFileSync(path.join(unzippedFolder, 'index.mjs'), 'utf8');
         expect(fileContent).toContain('@aws-sdk/client-neptune');
     });
+
+    test('Can query app sync API successfully', async () => {
+        const awsResources = JSON.parse(fs.readFileSync(path.join(outputFolderPath, 'AirportsJestSDKTest-resources.json'), 'utf8'));
+        const apiId = awsResources.AppSyncAPI;
+        const region = awsResources.region;
+        const results = await executeAppSyncQuery(apiId, 'query {getNodeContinents {code}}', {}, region);
+        const codes = results.data.getNodeContinents.map(continent => continent.code).sort();
+        expect(codes).toEqual(['AF', 'AN', 'AS', 'EU', 'NA', 'OC', 'SA']);
+    }, 600000);
 });